requirements.txt
distro==1.8.0
numpy==1.24.0
pandas==1.5.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
tabula-py==2.6.0
